Separate KeyboardLayout from Keyboard and add new config properties
Keyboard class contained huge code that is caused hard to read. The
logic separate KeyboardLayout from Keyboard class is separate view logic
from operational logic. So readable and well-defined code structure was
produced. 
  Switching between keyboard layouts was made by hard code so any change
from keyboard-config file could broke the operation. Now just create any
key with switch id then set its target property value with the layout id
and it will switch to the target keyboard layout. When keyboard was
shown, the initial layout was germanLayout. The reason is it was
hardcoded as default layout. Now initial property added to the
keyboard-config file so it is possible change initial layout. 

Change-Id: I3ba11ad89b1c69fde7ca337ccfa78a386cb04eb7