Convert the project into an npm module
Allow including the keyboard in an app as a module.

`npm install` skips git-ignored directories, including `dist`, which
must be included in the module. Add an .npmignore file that does not
contain this folder.

Relative paths to images are resolved relative to the app's root dir,
not the module's root dir. Add a prefix to image paths to make them
available in the app.

Make `tabris` a dev-dependency, as it should be provided by the app
that uses the module.

Change-Id: I2c810fd1f3734fafc8ba200ecaf1066ab27d98ff